move byte[] to tests
